requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
six==1.16.0
urllib3==1.26.9
